Rearranged unsafe logic + added awaken()
There was an unsafe path where changes were made to the
state and it was then possible to enter the slowpath.
I moved some code around to make it so that this
doesn't happen. Additionally, I added the check part
of awaken(), which will redirect to the slowpath if
it succeeds as to make this a more strict subset of
the signal slowpath. Added some extra asserts that
appear in the signal slowpath.

Signed-off-by: Alwin Joshy <a.joshy@student.unsw.edu.au>