Removed signal to higher prio path
After the discussion on the PR, it was decided that the best course
of action was to only keep the signal to low prio threads fastpath. I
removed the signal to higher prio fastpath and all the things it used
that are no longer necessary. I also fixed a cmake misconfiguration that
previously made it annoying to switch the signal fastpath on and off.
Also added #ifdefs to signal fastpath so it wouldn't stop other
configurations from compiling. Changed sched_enqueue to
sched_append as this makes more sense.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>